Add new dynamic client and fake, fix waitforagents missing in crd manifest
In order to implement a fake client (easily) it was necessary to
superceded the existing rest client with the dynamic client. This makes
it possible to create a fake client based simply on an object tracker.
The motivation for this is to get unit-testability for CRD generation in
KNE. It would have been wreckless to ignore when adding support for
WaitForAgents and ToggleOverrides over there.